+++ src/components/Menu.vue
@@ -29,11 +29,8 @@ export default {
       menu : [
         {label: this.$t('menu.games')},
         {label: this.$t('menu.questions'), to:'/questions'},
-        {label: this.$t('menu.logos'), to:'/logos'},
-        // {label: this.$t('menu.nutritions'), to:'/nutritions'},
         {},
         {label: this.$t('menu.manage'), type: 'header'},
-        {label: this.$t('menu.insights'), to:'/insights'},
         {label: this.$t('menu.settings'), to:'/settings'},
       ]
     };
