+++ src/const.js
@@ -1,4 +1,6 @@
-export const ROBOTOFF_API_URL = "https://robotoff.openfoodfacts.org/api/v1";
+// export const ROBOTOFF_API_URL = "https://robotoff.openfoodfacts.org/api/v1";
+export const ROBOTOFF_API_URL = "https://robotoff.wild31.com/api/v1";
+
 export const OFF_URL = "https://world.openfoodfacts.org";
 export const OFF_API_URL = `${OFF_URL}/api/v0`;
 export const OFF_IMAGE_URL = "https://static.openfoodfacts.org/images/products";
