--- src/data/nutritions.json
@@ -1,12 +1,0 @@
-{
-    "nutriment_energy-kj": null,
-    "nutriment_energy-kcal": null,
-    "nutriment_fat": "g",
-    "nutriment_saturated-fat": "g",
-    "nutriment_carbohydrates": "g",
-    "nutriment_sugars": "g",
-    "nutriment_proteins": "g",
-    "nutriment_fiber": "g",
-    "nutriment_salt": "g",
-    "nutriment_sodium": null
-}
