+++ src/main.js
@@ -5,13 +5,7 @@ import Viewer from 'v-viewer'
 import './semantic-ui-utils/main.css'
 import App from './App.vue'
 import QuestionView from './views/QuestionView.vue'
-import NutritionView from './views/NutritionView.vue'
-import TableAnnotation from './views/TableAnnotation.vue'
 import SettingsView from './views/SettingsView.vue'
-import InsightListView from './views/InsightListView.vue'
-import LogoSearchView from './views/LogoSearchView.vue'
-import LogoAnnotationView from './views/LogoAnnotationView.vue'
-import LogoUpdateView from './views/LogoUpdateView.vue'
 import 'viewerjs/dist/viewer.css'
 import SuiVue from 'semantic-ui-vue'
 import messages from './i18n/messages'
@@ -26,14 +20,8 @@ Vue.config.productionTip = false
 
 const routes = [
   { path: '/', redirect: '/questions' },
-  { path: '/insights', component: InsightListView },
   { path: '/questions', component: QuestionView },
-  { path: '/nutritions', component: NutritionView },
-  { path: '/table', component: TableAnnotation },
   { path: '/settings', component: SettingsView },
-  { path: '/logos', component: LogoAnnotationView },
-  { path: '/logos/search', component: LogoSearchView },
-  { path: '/logos/:id', component: LogoUpdateView },
 ];
 
 const router = new VueRouter({
