+++ src/views/QuestionView.vue
@@ -11,7 +11,7 @@
         >
           {{ $t("questions." + insightType) }}
         </div>
-        <div class="ui form">
+        <!-- <div class="ui form">
           <div class="ui icon input" id="value-tag-input">
             <input
               class="ui input"
@@ -28,7 +28,7 @@
             <input v-model="sortByPopularity" type="checkbox" name="sortBy" />
             <label>{{ $t("questions.popularity_sort") }}</label>
           </div>
-        </div>
+        </div> -->
         <div class="ui divider" />
         <div class="ui hidden divider"></div>
         <div v-if="currentQuestion">
@@ -111,94 +111,15 @@ import LoadingSpinner from "../components/LoadingSpinner";
 
 import AnnotationCounter from "../components/AnnotationCounter";
 
-const updateURLParam = (key, value) => {
-  const urlParams = new URLSearchParams(window.location.search);
-  urlParams.set(key, value);
-  const newRelativePathQuery =
-    window.location.pathname + "?" + urlParams.toString();
-  history.pushState(null, "", newRelativePathQuery);
-};
-
-const deleteURLParam = (key) => {
-  const urlParams = new URLSearchParams(window.location.search);
-
-  if (!urlParams.has(key)) {
-    return;
-  }
-
-  urlParams.delete(key);
-
-  const newRelativePathQuery =
-    window.location.pathname +
-    (urlParams.toString().length ? `?${urlParams.toString()}` : "");
-
-  history.pushState(null, "", newRelativePathQuery);
-};
-
-const getURLParam = (key) => {
-  const urlParams = new URLSearchParams(window.location.search);
-
-  if (!urlParams.has(key)) {
-    return "";
-  }
-
-  return urlParams.get(key);
-};
-
-const NO_QUESTION_LEFT = "NO_QUESTION_LEFT";
-
-const insightTypesNames = {
-  label: "label",
-  category: "category",
-  brand: "brand",
-  product_weight: "product weight",
-};
-
-const randomInsightTypeChoices = ["label", "category", "brand"];
-
-const getRandomInsightType = () =>
-  randomInsightTypeChoices[
-    Math.floor(randomInsightTypeChoices.length * Math.random())
-  ];
-
-const getInitialInsightType = () => {
-  const urlParamValue = getURLParam("type");
-
-  if (urlParamValue.length) {
-    return urlParamValue;
-  }
-
-  return getRandomInsightType();
-};
-
-const reformatTagMapping = {
-  " ": "-",
-  "'": "-",
-  "&": "",
-  à: "a",
-  â: "a",
-  ä: "a",
-  é: "e",
-  è: "e",
-  ê: "e",
-  ë: "e",
-  î: "i",
-  ï: "i",
-  ô: "o",
-  ö: "o",
-  û: "u",
-  ù: "u",
-  ü: "u",
-};
-
-const reformatValueTag = (value) => {
-  let output = value.trim().toLowerCase();
-  for (const [search, replace] of Object.entries(reformatTagMapping)) {
-    output = output.replace(new RegExp(search, "g"), replace);
-  }
-  output = output.replace(/-{2,}/g, "-");
-  return output;
-};
+import {
+  updateURLParam,
+  deleteURLParam,
+  getURLParam,
+  NO_QUESTION_LEFT,
+  insightTypesNames,
+  getInitialInsightType,
+  reformatValueTag
+} from "../utils/utilsQuestionView"
 
 export default {
   name: "QuestionView",
